create ImageProcessor component
and refactor stage1 to use it
